change dtype

+++ ML/Darts/Utils/preprocessing.py
@@ -105,9 +105,9 @@ def load_historical_data(data:Historical, period:int) -> TimeSeries:
         "timestamp": [datetime.fromtimestamp(value[0]) for value in data.data["data"]["result"][0]["values"]],
         "value": [float(value[1]) for value in data.data["data"]["result"][0]["values"]]
     }
-    df = pd.DataFrame(series).astype(np.float32)
+    df = pd.DataFrame(series)
     df['timestamp'] = pd.to_datetime(df['timestamp']).dt.tz_localize(None)
-    final_series = TimeSeries.from_dataframe(df, time_col='timestamp', value_cols='value', fill_missing_dates=True, freq=f"{period}ms")
+    final_series = TimeSeries.from_dataframe(df, time_col='timestamp', value_cols='value', fill_missing_dates=True, freq=f"{period}ms").astype(np.float32)
     return final_series
